cgi: Add Python version, page creation time and make CGI PEP8 compliant
This commit adds creation time and Python version to all CGI pages and
formats CGI Python scripts with respect to the PEP8 standard using
autopep8 tool.

Creation time is added in the footer part (right), since it is the most
good-looking part of the whole CGI to add such information.  Space (left
and center) is also reserved for possible future informations to show.

Closes #242

Change-Id: I1f4d488e41572388f00d19c03b18b71c5cba3587