cgi: Fix bugs in CGI after upgrade to Python3
None is replaced by LessThanComparableNone class's instance when
used as an element of a tuple which has to be compared with other
tuples when sorting them.
This is because of a change in Python3 that None cannot be used in
ordering comparison. HTML generation code in CGI project uses the fact
that in Python2 None is considered less than any value of
a non-NoneType type in ordering comparisons.
LessThanComparableNone type has exactly this property. That is
a special null-like value that can take part in ordering comparisons.

Change-Id: I0e4bb4cb0707c95720ebf457c5e1d278c3e7b970